ppl9: reading of very long lines allowed
git-svn-id: svn://127.0.0.1/pplsvn@1288 8e462c0f-ee70-4953-a96e-5db927b55778

+++ src/expressions/expCompile.h
@@ -30,7 +30,7 @@ typedef struct pplExpr {
 
 typedef struct pplTokenCode {
   unsigned char  state, opcode, precedence;
-  unsigned short depth;
+  unsigned int   depth;
  } pplTokenCode;
 
 typedef struct pplExprPStack {
